Split run function into separate files for OS builds.
