+++ cmd/util_windows.go
@@ -1,0 +1,28 @@
+//go:build windows
+
+package cmd
+
+import (
+	"fmt"
+	"io"
+	"os"
+	"os/exec"
+	"syscall"
+
+	"github.com/knqyf263/pet/config"
+)
+
+func run(command string, r io.Reader, w io.Writer) error {
+	var cmd *exec.Cmd
+	if len(config.Conf.General.Cmd) > 0 {
+		line := append(config.Conf.General.Cmd, command)
+		cmd = exec.Command(line[0], line[1:]...)
+	} else {
+		cmd = exec.Command("cmd.exe")
+		cmd.SysProcAttr = &syscall.SysProcAttr{CmdLine: fmt.Sprintf("/c \"%s\"", command)}
+	}
+	cmd.Stderr = os.Stderr
+	cmd.Stdout = w
+	cmd.Stdin = r
+	return cmd.Run()
+}
